Reorganizing to allow certain libraries to only access tabular plots
