+++ hvplot/pandas.py
@@ -1,3 +1,18 @@
-from . import patch
+def patch(name='hvplot', extension='bokeh', logo=False):
+    from . import plotting
+    from . import post_patch
 
-patch('pandas', extension='bokeh')
+    try:
+        import pandas as pd
+    except:
+        raise ImportError('Could not patch plotting API onto pandas. '
+                          'Pandas could not be imported.')
+    _patch_plot = lambda self: plotting.hvPlotTab(self)
+    _patch_plot.__doc__ = plotting.hvPlotTab.__call__.__doc__
+    patch_property = property(_patch_plot)
+    setattr(pd.DataFrame, name, patch_property)
+    setattr(pd.Series, name, patch_property)
+
+    post_patch(extension, logo)
+
+patch()
