+++ hvplot/tests/testpatch.py
@@ -6,23 +6,23 @@ from unittest import TestCase, SkipTest
 
 import numpy as np
 
-from hvplot import patch, hvPlot
+from hvplot.plotting import hvPlot, hvPlotTab, hvPlotGridded
 
 
 class TestPatchPandas(TestCase):
 
     def setUp(self):
-        patch('pandas')
+        import hvplot.pandas
 
     def test_pandas_series_patched(self):
         import pandas as pd
         series = pd.Series([0, 1, 2])
-        self.assertIsInstance(series.hvplot, hvPlot)
+        self.assertIsInstance(series.hvplot, hvPlotTab)
 
     def test_pandas_dataframe_patched(self):
         import pandas as pd
         df = pd.DataFrame([[1, 2], [3, 4], [5, 6]], columns=['x', 'y'])
-        self.assertIsInstance(df.hvplot, hvPlot)
+        self.assertIsInstance(df.hvplot, hvPlotTab)
 
 
 class TestPatchDask(TestCase):
@@ -32,21 +32,21 @@ class TestPatchDask(TestCase):
             import dask.dataframe as dd # noqa
         except:
             raise SkipTest('Dask not available')
-        patch('dask')
+        import hvplot.dask
 
     def test_dask_series_patched(self):
         import pandas as pd
         import dask.dataframe as dd
         series = pd.Series([0, 1, 2])
         dseries = dd.from_pandas(series, 2)
-        self.assertIsInstance(dseries.hvplot, hvPlot)
+        self.assertIsInstance(dseries.hvplot, hvPlotTab)
 
     def test_dask_dataframe_patched(self):
         import pandas as pd
         import dask.dataframe as dd
         df = pd.DataFrame([[1, 2], [3, 4], [5, 6]], columns=['x', 'y'])
         ddf = dd.from_pandas(df, 2)
-        self.assertIsInstance(ddf.hvplot, hvPlot)
+        self.assertIsInstance(ddf.hvplot, hvPlotTab)
 
 
 class TestPatchXArray(TestCase):
@@ -62,15 +62,15 @@ class TestPatchXArray(TestCase):
         import xarray as xr
         array = np.random.rand(100, 100)
         xr_array = xr.DataArray(array, coords={'x': range(100), 'y': range(100)}, dims=('y', 'x'))
-        self.assertIsInstance(xr_array.hvplot, hvPlot)
+        self.assertIsInstance(xr_array.hvplot, hvPlotGridded)
 
     def test_xarray_dataset_patched(self):
         import xarray as xr
         array = np.random.rand(100, 100)
         xr_array = xr.DataArray(array, coords={'x': range(100), 'y': range(100)}, dims=('y', 'x'))
         xr_ds = xr.Dataset({'z': xr_array})
-        self.assertIsInstance(xr_ds.hvplot, hvPlot)
-        
+        self.assertIsInstance(xr_ds.hvplot, hvPlotGridded)
+
 
 class TestPatchStreamz(TestCase):
 
@@ -79,24 +79,24 @@ class TestPatchStreamz(TestCase):
             import streamz # noqa
         except:
             raise SkipTest('streamz not available')
-        patch('streamz')
+        import hvplot.streamz
 
     def test_streamz_dataframe_patched(self):
         from streamz.dataframe import Random
         random_df = Random()
-        self.assertIsInstance(random_df.hvplot, hvPlot)
+        self.assertIsInstance(random_df.hvplot, hvPlotTab)
 
     def test_streamz_series_patched(self):
         from streamz.dataframe import Random
         random_df = Random()
-        self.assertIsInstance(random_df.x.hvplot, hvPlot)
+        self.assertIsInstance(random_df.x.hvplot, hvPlotTab)
 
     def test_streamz_dataframes_patched(self):
         from streamz.dataframe import Random
         random_df = Random()
-        self.assertIsInstance(random_df.groupby('x').sum().hvplot, hvPlot)
+        self.assertIsInstance(random_df.groupby('x').sum().hvplot, hvPlotTab)
 
     def test_streamz_seriess_patched(self):
         from streamz.dataframe import Random
         random_df = Random()
-        self.assertIsInstance(random_df.groupby('x').sum().y.hvplot, hvPlot)
+        self.assertIsInstance(random_df.groupby('x').sum().y.hvplot, hvPlotTab)
